Refactor ComfyUI helpers
